The beginning of the translation. I'd like to point out a few code flaws.
First, the use of Tangerine font is VERY doubtful. Is it really need to
download a FONT just for a ONE symbol? Of course, this is all about HTML
5 and the future, but Tangerine contains only Latin symbols, so I can't
use it anywhere else. We're still in 2010, not in 2013 or something!

Second, the FUDGE grades of skills are hardly coded. This is really bad
when comes to translation.

Third, there is no easy way to create a title before title. Some screen
of "choose-your-language". The title screen accepts only clicks, so no
choices - and the title screen already must be translated! Obviously,
there's too late for "choose-your language" after the first Situation is
loaded. It's possible to load the game script dynamically, but in this
language, not in this engine.

Also I had to modify the style because it has no h3 rule for title. And
of course, text in HTML is also hard coded.