Resolve #25
When setQuality is called on a quality that isn't defined,
the quality is set in the character (as the documentation says), but
Undum no longer crashes when trying to display the quality (it fails
silently). On the other hand, calling addQualityBlock for a quality
without a definition throws a more sensible error message than a
complaint about trying to access 'null'.